[wip] restructure code
