+++ src/lib/auth.ts
@@ -1,0 +1,77 @@
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import { compare } from 'bcryptjs';
+import { AuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import GithubProvider from 'next-auth/providers/github';
+import GoogleProvider from 'next-auth/providers/google';
+
+import prisma from './prisma';
+
+const authOptions: AuthOptions = {
+  pages: {
+    /* override signIn, signOut, error page urls here... */
+  },
+  adapter: PrismaAdapter(prisma),
+  session: {
+    strategy: 'jwt',
+  },
+  callbacks: {
+    /* define callback options here */
+  },
+  providers: [
+    CredentialsProvider({
+      name: 'credentials',
+      type: 'credentials',
+      credentials: {
+        username: {
+          label: 'Username',
+          type: 'text',
+          placeholder: 'Enter your username',
+        },
+        password: {
+          label: 'Password',
+          type: 'password',
+          placeholder: 'Enter your password',
+        },
+      },
+      async authorize(credentials, req) {
+        if (!credentials?.username || !credentials.password) {
+          return null;
+        }
+
+        const user = await prisma.user.findUnique({
+          where: {
+            username: credentials.username,
+          },
+        });
+
+        if (
+          !user ||
+          !user.password ||
+          !(await compare(credentials.password, user.password))
+        ) {
+          return null;
+        }
+
+        return {
+          id: user.id,
+          username: user.username,
+          name: user.name,
+          email: user.email,
+          emailVerified: user.emailVerified,
+          image: user.image,
+        };
+      },
+    }),
+    GithubProvider({
+      clientId: process.env.GITHUB_CLIENT_ID as string,
+      clientSecret: process.env.GITHUB_CLIENT_SECRET as string,
+    }),
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID as string,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+    }),
+  ],
+};
+
+export default authOptions;
